--- src/auth/application/authorize-oauth/dto/authorize-oauth.request.dto.ts
@@ -1,8 +1,0 @@
-import { IsEnum, IsNotEmpty } from 'class-validator';
-import { OAuthProviderType } from 'src/auth/domain/value-object/oauth-provider.enum';
-
-export class AuthorizeOAuthRequestDto {
-  @IsEnum(OAuthProviderType)
-  @IsNotEmpty()
-  oAuthProviderType: OAuthProviderType;
-}
