+++ src/auth/auth.module.ts
@@ -1,44 +1,11 @@
 import { Module } from '@nestjs/common';
-import { AUTH_REPOSITORY } from './domain/repository/auth.repository';
-import { AuthRepositoryImpl } from './infrastructure/repository/auth.repository.impl';
-import { AuthController } from './presentation/auth.controller';
-import { MikroOrmModule } from '@mikro-orm/nestjs';
-import { AuthEntity } from './infrastructure/orm-entity/auth.entity';
-import { OAuthLoginUseCase } from './application/oauth-login/oauth-login.use-case';
-import { OAuthProviderFactory } from './infrastructure/factory/oauth-provider.factory';
-import { KakaoOAuthProvider } from './infrastructure/provider/kakao.provider';
-import { AuthorizeOAuthUseCase } from './application/authorize-oauth/authorize-oauth.use-case';
-import { JwtProvider } from './infrastructure/provider/jwt.provider';
 import { JwtModule } from '@nestjs/jwt';
-import { JwtAccessStrategy } from './infrastructure/strategy/jwt-access.strategy';
-import { JwtRefreshStrategy } from './infrastructure/strategy/jwt-refresh.strategy';
 import { PassportModule } from '@nestjs/passport';
-import { RenewTokenUseCase } from './application/renew-token/renew-token.use-case';
-import { LogoutUseCase } from './application/logout/logout.use-case';
 import { UserCommandModule } from 'src/user/command/user.command.module';
-import { UserEntity } from 'src/user/command/infrastructure/user.entity';
-
-const useCases = [OAuthLoginUseCase, AuthorizeOAuthUseCase, RenewTokenUseCase, LogoutUseCase];
+import { AuthCommandModule } from './command/auth.command.module';
+import { AuthQueryModule } from './query/auth.query.module';
 
 @Module({
-  controllers: [AuthController],
-  imports: [
-    JwtModule.register({}),
-    MikroOrmModule.forFeature([AuthEntity, UserEntity]),
-    PassportModule,
-    UserCommandModule,
-  ],
-  providers: [
-    ...useCases,
-    {
-      provide: AUTH_REPOSITORY,
-      useClass: AuthRepositoryImpl,
-    },
-    OAuthProviderFactory,
-    KakaoOAuthProvider,
-    JwtProvider,
-    JwtAccessStrategy,
-    JwtRefreshStrategy,
-  ],
+  imports: [JwtModule.register({}), AuthCommandModule, AuthQueryModule, PassportModule, UserCommandModule],
 })
 export class AuthModule {}
