+++ src/auth/query/application/authorize-oauth/dto/authorize-oauth.request.dto.ts
@@ -1,0 +1,8 @@
+import { IsEnum, IsNotEmpty } from 'class-validator';
+import { OAuthProviderType } from 'src/auth/command/domain/value-object/oauth-provider.enum';
+
+export class AuthorizeOAuthRequestDto {
+  @IsEnum(OAuthProviderType)
+  @IsNotEmpty()
+  oAuthProviderType: OAuthProviderType;
+}
