+++ src/auth/query/application/authorize-oauth/dto/authorize-oauth.response.dto.ts
